Active on stroj

+++ slc_week.py
@@ -28,6 +28,7 @@ vii) Create mean weekly mosaic from all resampled images for that week
 
 import glob
 import os
+import rasterio
 import time
 from datetime import datetime, timedelta
 from shutil import rmtree
@@ -37,6 +38,7 @@ from osgeo import gdal
 from scipy.ndimage import binary_dilation
 
 from composite_dask import composite
+from tif2jpg import tif2jpg
 
 
 class WeekList:
@@ -153,11 +155,13 @@ def find_individual_images(list_of_days, src, direction, dt):
     return out_image_with_list
 
 
-def pre_process_bursts(bursts_list, polarity, folder_pth):
+def pre_process_bursts(bursts_list, polarity, folder_pth, bbox=None):
     """Prepares input images (bursts) for processing and returns list of paths.
 
         For each burst:
         1)  Sets path to pre-processed image
+        1a) Skips if image is out of bounds [time saving]
+        1b) New output bounds if only partial overlap [time saving]
         2)  Deal with nodata (all nodata should be 0)
         3)  Test if there are any "nodata stripes" across the image
         3a) If yes, then first erode by a couple pixels
@@ -168,65 +172,93 @@ def pre_process_bursts(bursts_list, polarity, folder_pth):
     for i, burst in enumerate(bursts_list):
         print(f"{i+1}", end="")
 
-        # Store paths of output, so they can be used in the nex step
+        # Determine full file name
         p = os.path.join(burst, f"*{polarity}*.img")
         burst_file = glob.glob(p)[0]
-        image_name = f"{i:02d}_" + os.path.basename(burst_file)[:-4] + ".tif"
-        out_burst = os.path.join(folder_pth, image_name)
-        paths.append(out_burst)
 
-        # Set all nodata to 0 (there are both 0 and nan values present)
-        ds = gdal.Warp(out_burst, burst_file, srcNodata=np.nan, dstNodata=0, multithread=True)
-        ds = None
+        # Check if extents overlap with AOI (if bbox was assigned)
+        if bbox:
+            with rasterio.open(burst_file) as rio:
+                # Extents of the burst
+                bbr = rio.bounds
+            is_overlapping = (
+                    bbr.left < bbox[2] and bbox[0] < bbr.right and
+                    bbr.bottom < bbox[3] and bbox[1] < bbr.top)
+            out_bounds = list(bbr)
+            if bbr.left < bbox[0]:
+                out_bounds[0] = bbox[0]
+            if bbr.bottom < bbox[1]:
+                out_bounds[1] = bbox[1]
+            if bbr.right > bbox[2]:
+                out_bounds[2] = bbox[2]
+            if bbr.top > bbox[1]:
+                out_bounds[3] = bbox[3]
+        else:
+            is_overlapping = True
+            out_bounds = None
 
-        # Open raster for further processing
-        r = gdal.Open(out_burst, gdal.GA_Update)  # gdal.GA_Update: save output to the source file
-        raster_arr = r.GetRasterBand(1).ReadAsArray()
-        warp_iter = 10  # By default erode edges by 10 pixels
+        if is_overlapping:
+            # Store paths of output, so they can be used in the nex step
+            image_name = f"{i:02d}_" + os.path.basename(burst_file)[:-4] + ".tif"
+            out_burst = os.path.join(folder_pth, image_name)
+            paths.append(out_burst)
 
-        # Test for nodata "stripes" (by reading a random column and counting nodata intervals
-        test_column = raster_arr[:, raster_arr.shape[1] // 2]
-        a_zeros = np.where(test_column == 0)[0]
-        a_groups = np.split(a_zeros, np.where(np.diff(a_zeros) != 1)[0] + 1)
-
-        # More than 2 nodata intervals means there is a nodata stripe present
-        if len(a_groups) > 2:
-            # Find the width of nodata pixels
-            px_wid = len(a_groups[1])
-            # First erode by 3 pixels to remove dark pixels on edge of strip
-            nodata_mask = raster_arr == 0
-            dilated_mask = binary_dilation(nodata_mask, iterations=3)
-            raster_arr[dilated_mask] = 0
-
-            # Prepare raster band to be used by fill nodata
-            r.GetRasterBand(1).WriteArray(raster_arr)
-            raster_bnd = r.GetRasterBand(1)
-            # Preform fill nodata on the array
-            ds = gdal.FillNodata(raster_bnd, maskBand=None,
-                                 maxSearchDist=px_wid+3, smoothingIterations=0,
-                                 callback=None)
+            # Set all nodata to 0 (there are both 0 and nan values present)
+            ds = gdal.Warp(out_burst, burst_file, srcNodata=np.nan,
+                           dstNodata=0, multithread=True,
+                           outputBounds=out_bounds)
             ds = None
 
-            # Open band as array to be dilated in the next step
-            raster_arr = raster_bnd.ReadAsArray()
-            # Also erode pixels that were added with FillNodata
-            warp_iter += px_wid
+            # Open raster for further processing
+            r = gdal.Open(out_burst, gdal.GA_Update)  # gdal.GA_Update: save output to the source file
+            raster_arr = r.GetRasterBand(1).ReadAsArray()
+            warp_iter = 10  # By default erode edges by 10 pixels
 
-        # Remove dark pixels on the edge of each raster
-        nodata_mask = raster_arr == 0
-        dilated_mask = binary_dilation(nodata_mask, iterations=warp_iter)
-        raster_arr[dilated_mask] = 0
+            # Test for nodata "stripes" (by reading a random column and counting nodata intervals
+            test_column = raster_arr[:, raster_arr.shape[1] // 2]
+            a_zeros = np.where(test_column == 0)[0]
+            a_groups = np.split(a_zeros, np.where(np.diff(a_zeros) != 1)[0] + 1)
 
-        r.GetRasterBand(1).WriteArray(raster_arr)
-        # noinspection PyUnusedLocal
-        r = None
+            # More than 2 nodata intervals means there is a nodata stripe present
+            if len(a_groups) > 2:
+                # Find the width of nodata pixels
+                px_wid = len(a_groups[1])
+                # First erode by 3 pixels to remove dark pixels on edge of strip
+                nodata_mask = raster_arr == 0
+                dilated_mask = binary_dilation(nodata_mask, iterations=3)
+                raster_arr[dilated_mask] = 0
 
-        print(f"X ", end="")
+                # Prepare raster band to be used by fill nodata
+                r.GetRasterBand(1).WriteArray(raster_arr)
+                raster_bnd = r.GetRasterBand(1)
+                # Preform fill nodata on the array
+                ds = gdal.FillNodata(raster_bnd, maskBand=None,
+                                     maxSearchDist=px_wid+3, smoothingIterations=0,
+                                     callback=None)
+                ds = None
 
+                # Open band as array to be dilated in the next step
+                raster_arr = raster_bnd.ReadAsArray()
+                # Also erode pixels that were added with FillNodata
+                warp_iter += px_wid
+
+            # Remove dark pixels on the edge of each raster
+            nodata_mask = raster_arr == 0
+            dilated_mask = binary_dilation(nodata_mask, iterations=warp_iter)
+            raster_arr[dilated_mask] = 0
+
+            r.GetRasterBand(1).WriteArray(raster_arr)
+            # noinspection PyUnusedLocal
+            r = None
+
+            print(f"X ", end="")
+        else:
+            # Message next to the burst number if image is out of bounds
+            print(f":n/a ", end="")
     return paths
 
 
-def make_individual_rasters(to_aggregate, bbox, direct, polar, tmp_folder):
+def make_individual_rasters(to_aggregate, direct, polar, tmp_folder, bbox=None):
     """Prepares all individual products from one week for compositing.
 
     Parameters
@@ -235,14 +267,14 @@ def make_individual_rasters(to_aggregate, bbox, direct, polar, tmp_folder):
         List of tuples, with first element being a product and second a list of
         paths to all corresponding sub-products (folders containing *VV.img anf
         *VH.img files).
-    bbox : list
-        Output extents in the [xmin, ymin, xmax, ymax] format
     direct : str
         Orbit direction, either ASC for Ascending or DES for Descending.
     polar : str
         Polarity, either VV or VH for S-1 SLC products
     tmp_folder : str
         Path for saving outputs.
+    bbox : list
+        Output extents in the [x_min, y_min, x_max, y_max] format
 
     Returns
     -------
@@ -272,35 +304,39 @@ def make_individual_rasters(to_aggregate, bbox, direct, polar, tmp_folder):
         # Pre-process "bursts" for warping into a single image
         # to_be_warped is a LIST OF PATHS to individual product folders
         print(f"        - consists of {len(bursts)} bursts\n        ", end="")
-        to_be_warped = pre_process_bursts(bursts, polar, tmp_folder)
+        to_be_warped = pre_process_bursts(bursts, polar, tmp_folder, bbox=bbox)
 
-        # WARP BURSTS INTO SINGLE IMAGE
-        out_image = os.path.join(tmp_folder, product + f"_{direct}_{polar}.tif")
-        final_paths.append(out_image)
-        print(f"\n        - warping into a single image")
+        if to_be_warped:
+            # WARP BURSTS INTO SINGLE IMAGE
+            out_image = os.path.join(tmp_folder, product + f"_{direct}_{polar}.tif")
+            final_paths.append(out_image)
+            print(f"\n        - warping into a single image")
 
-        # Resample to 10m using bilinear interpolation and align pixels to grid
-        # Also crop to extents - all files should have the same extents (outputBounds)
-        ds = gdal.Warp(out_image, to_be_warped,
-                       xRes=10, yRes=10,
-                       dstNodata=0, targetAlignedPixels=True,
-                       resampleAlg=gdal.gdalconst.GRA_Bilinear,
-                       multithread=True, outputBounds=bbox,
-                       options=['TILED=YES', 'BLOCKXSIZE=512', 'BLOCKYSIZE=512'])
-        ds = None
+            # Resample to 10m using bilinear interpolation and align pixels to grid
+            # Also crop to extents - all files should have the same extents (outputBounds)
+            ds = gdal.Warp(out_image, to_be_warped,
+                           xRes=10, yRes=10,
+                           dstNodata=0, targetAlignedPixels=True,
+                           resampleAlg=gdal.gdalconst.GRA_Bilinear,
+                           multithread=True, outputBounds=bbox,
+                           options=['TILED=YES', 'BLOCKXSIZE=512', 'BLOCKYSIZE=512'])
+            ds = None
 
-        # REMOVE TEMPORARY FILES ("bursts")
-        for file in to_be_warped:
-            os.remove(file)
+            # REMOVE TEMPORARY FILES ("bursts")
+            for file in to_be_warped:
+                os.remove(file)
 
-        tta1 = time.time() - tta1
-        print(f"        [Time (individual image): {tta1:.2f} sec.]")
+            tta1 = time.time() - tta1
+            print(f"        [Time (individual image): {tta1:.2f} sec.]")
+        else:
+            print(f"\n        - no images inside bounds... SKIPPING")
 
     return final_paths
 
 
 def loop_weeks(dt_start, dt_end, dt_step, bbox, data_type,
-               src_folder, save_loc, combinations=None):
+               src_folder, save_loc,
+               combinations=None, country_border=None):
     # Create object for finding time intervals for processing
     my_weeks = WeekList(dt_start, dt_end, dt_step)
 
@@ -369,8 +405,8 @@ def loop_weeks(dt_start, dt_end, dt_step, bbox, data_type,
 
             # ==================================================================
             # PROCESS INDIVIDUAL IMAGES
-            paths_for_composite = make_individual_rasters(to_aggregate, bbox, direct,
-                                                          polar, tmp_f)
+            paths_for_composite = make_individual_rasters(to_aggregate, direct,
+                                                          polar, tmp_f, bbox=bbox)
             t_combo = time.time() - t_combo
             print(f"\n  Finished combo {direct} {polar} in {t_combo:.2f} sec.")
 
@@ -382,8 +418,10 @@ def loop_weeks(dt_start, dt_end, dt_step, bbox, data_type,
                 tww = this_week["week"]
                 composite_name = f"{diw[0]}_{diw[-1]}_weekly_SLC_{data_type}" \
                                  f"_{direct}_{polar}_yr{diw[0][2:4]}wk{tww:02}"
-                composite(paths_for_composite, week_path, composite_name,
-                          method="mean", dt=data_type)
+                tif = composite(paths_for_composite, week_path, composite_name,
+                                method="mean", dt=data_type)
+                # CREATE JPG PREVIEW
+                tif2jpg(tif, country_border)
                 tta2 = time.time() - tta2
                 print(f"#\n# Time (composite + preview file): {tta2:.2f} sec.\n")
             else:
@@ -415,7 +453,9 @@ if __name__ == "__main__":
     # This bbox is outline of SLO with 5 km buffer in EPSG:32633
     # bbox = [368930, 5024780, 627570, 5197200]
     # bbox for Netherlands in EPSG:28992
-    in_bbox = [0, 305400, 287100, 625100]
+    in_bbox = [0, 305400, 287100, 625100]  # NL full
+    # in_bbox = [90000, 311000, 140000, 554591]  # NL partially out of bounds
+    # in_bbox = [387200, 740000, 400000, 840000]  # NL completely out of bounds
 
     in_comb = None
     # combinations = [
@@ -431,7 +471,7 @@ if __name__ == "__main__":
     # in_src = "r:\\Sentinel-1_SLC_products_SI_coherence_13.91m_UTM33N"
     # in_src = "r:\\Sentinel-1_SLC_products_SI_sigma_10m_UTM33N"
     # in_src = "r:\\Sentinel-1_SLC_products_AiTLAS_NL_coh_10m_Amersfoort"
-    in_src = "q:\\_S1_SLC_products_NL_sigma_10m_Amersfoort"
+    in_src = "q:\\_S1_SLC_products_NL_sig_10m_Amersfoort"
 
     # Save location
     in_save = "d:\\aitlas_slc_test_NL_2"
